+++ package.json
@@ -33,8 +33,8 @@
     "serve-rel-sql-playground": "cd apps/relSQLPlayground/public && http-server -c-1",
     "build-vscode-extension": "esbuild apps/vscodeExtension/extension.ts --bundle --sourcemap --platform=node --outfile=apps/vscodeExtension/out/extension.js --external:vscode --loader:.dl=text --loader:.grammar=text --loader:.html=text && esbuild apps/vscodeExtension/webView.tsx --bundle --sourcemap '--define:process.env.NODE_ENV=\"development\"' --outfile=apps/vscodeExtension/out/webView.js --external:path --external:fs --external:crypto --loader:.dl=text --loader:.grammar=text --loader:.html=text",
     "package-vscode-extension": "cd apps/vscodeExtension && vsce package",
-    "build-datalog-parser": "ts-node languageWorkbench/parserlib/gen/extractorCLI.ts languageWorkbench/languages/dl/dl.grammar languageWorkbench/languages/dl/parser.ts ../../parserlib DL ws && prettier --write languageWorkbench/languages/dl/parser.ts",
-    "build-sql-parser": "ts-node languageWorkbench/parserlib/gen/extractorCLI.ts languageWorkbench/languages/sql/sql.grammar languageWorkbench/languages/sql/parser.ts ../../parserlib SQL ws && prettier --write languageWorkbench/languages/sql/parser.ts"
+    "build-datalog-parser": "ts-node languageWorkbench/parserlib/gen/cli.ts languageWorkbench/languages/dl/dl.grammar languageWorkbench/languages/dl/parser.ts ../../parserlib DL ws && prettier --write languageWorkbench/languages/dl/parser.ts",
+    "build-sql-parser": "ts-node languageWorkbench/parserlib/gen/cli.ts languageWorkbench/languages/sql/sql.grammar languageWorkbench/languages/sql/parser.ts ../../parserlib SQL ws && prettier --write languageWorkbench/languages/sql/parser.ts"
   },
   "devDependencies": {
     "@types/astring": "^1.7.0",
